// src/api/datasets_api.ts
export interface components {
  schemas: {
// ...
    DisplaySettings: {
      contrast_limits: components["schemas"]["ContrastLimits"];
      /**
       * Color
       * Format: color
       */
      color?: string;
      /** Invert Lut */
      invert_lut: boolean;
    };
// ...
    SampleTypeEnum: "scalar" | "label";
// ...
    Volume: {
      /** Name */
      name: string;
      /** Description */
      description: string;
      /** Url */
      url: string;
      format: components["schemas"]["ArrayContainerTypeEnum"];
      transform: components["schemas"]["SpatialTransform"];
      sample_type: components["schemas"]["SampleTypeEnum"];
      content_type: components["schemas"]["ContentTypeEnum"];
      display_settings: components["schemas"]["DisplaySettings"];
      /** Subsources */
      subsources: components["schemas"]["Mesh"][];
    };
  };
}